// src/app/modules/busqueda-pagares/interfaces/togas.interface.ts
// Generated by https://quicktype.io



export interface Alumnos {
  index?:               number;
  idalumno?:            string;
  idperson:            string;
  idGrado:             string;
  idperiodotitulacion: string;
  carrera:             string;
  login:               string;
  nombre:              string;
  fechaInicio?:        string;
  fechaFin?:           string;
  idIest?:             string;
}

export interface AlumnoWebList {
  idAlumno: number,
  abrCarrera: string,
  carrera: string,
  grado: string,
  idGrado: number,
  idTest: number,
  index?: number,
  nombre: string,
}
export interface togasPorFechaRequest{
  idAlumno: string,
  idcaja: string,
  idgrado: string,
  fechaInicio: string,
  fechaFin: string,
}
export interface Grados {
  id:    number;
  grado: string;
}

export interface ValidacionPago{
  tieneRecibo: boolean,
  mensaje: string,
  generar?: string
}